requirements.txt
google-generativeai
PyPDF2
python-docx
python-dotenv
tk